requirements.txt
langchain
langchain-openai
langchain-community
pdfplumber
faiss-cpu
streamlit
python-dotenv
tiktoken
langchain-groq